Enable network API for recipe nutrition and enable search.
